move corefinement specific code from PMP to Operations_on_polyhedra
