+++ Operations_on_polyhedra/include/CGAL/internal/corefinement/Polyhedra_output_builder.h
@@ -29,6 +29,7 @@
 #include <CGAL/Polygon_mesh_processing/orientation.h>
 #include <CGAL/Side_of_triangle_mesh.h>
 #include <CGAL/internal/corefinement/predicates.h>
+#include <CGAL/internal/corefinement/connected_components.h>
 #include <CGAL/internal/corefinement/utils.h>
 #include <CGAL/array.h>
